Add the "png-patterns" directory.
This directory contains sample patterns in the "disks-and-lines"
diagram format. Right now the Havana pattern contains a discontinuity
at the end because it only covers a quarter of the pattern, but
eventually I'll teach "png2js" how to patch up that sort of thing.

To animate one of these patterns with png2js, run these two commands:
  png2js png-patterns/PatternName/*.png > scratch/e3.js
  firefox scratch/e2.html