Replace (fx,fy) with a radian measure (fa).
This makes for less user-friendly Javascript, since you can't
any longer express "Alice is facing Bob" by setting Alice.(fx,fy)
equal to Bob.(x,y). However, I think I'm moving toward the idea
of having PNG be the primary input format, so it's not a high
priority to have the Javascript be readable or editable by humans.
And having a single radian measure instead of two Cartesian
coordinates does simplify a lot of the math --- it removes an
arctangent or two in some places, for example.