Checkpoint on --join= work.
It seems to work, except for deduce_center_of_rotation(), which
still needs the mathy stuff fleshed out. You should be able to
run the following commands and get smooth animation:

    ./png2js --join=120 ../png-patterns/5pHavana/5pHavana10[123456].png
    ./png2js --join=0 ../png-patterns/5pHavana/5pHavana1* ../png-patterns/5pHavana/5pHavana201.png
    ./png2js ../png-patterns/5pHavana/*.png

Notice that when "--join=" is passed, the program expects that
the final frame will be an exact match to the first frame, except
for rotation by JOINING_SPIN degrees and permuting of the jugglers'
roles.  However, when "--join=" is *not* passed, the program expects
that the final frame will *not* be an exact match; if it is, you'll
get a hiccup in the animation because two frames in a row will be
identical.