Add Morgan's Reel to png-patterns/.
This is a "migrating" pattern, i.e., 09.png is the same as 01.png
except for the permutation of jugglers' roles AND a
sliding-but-not-rotating translation of everyone's position.
This kind of thing isn't supported by png2js yet, but it should
be supported soon.

Also notice that 05.png is the same as 01.png except for permutation
AND a flip over the y-axis. Again, png2js doesn't support this yet,
but it should be coming soon.