Separate the "shape" logic from the "representation" logic.
process_frame() has to answer two questions about each shape
returned from find_shapes_in_image():

    - What shape is this? (Circle, line segment,...)

    - What does this kind of shape represent in terms of
      the pattern itself? (Juggler, exchange-pass,...)

The old process_frame() mixed the two types of logic together,
and so it was kind of a mess --- in particular, the "what shape
is this" logic was very sketchy and fragile.

The new process_frame() shoves all of the former logic into a
new routine called copy_and_classify(), which ensures that
find_shapes_in_image() returns a vector of Shapes which have
already been completely classified as Circles, LineSegments,
or UnknownShapes. process_frame() itself can now concentrate on
the latter logic, and there is a logical place to put additional
code if we decide e.g. that the current definition of "line
segment" is too loose.