Recognize arrows as ooglewise passes; debug line-segment directions.
When DEBUG_IMAGES is #defined nonzero, we'll output an image named
"debug-lines.png" containing the current frame plus light-green line
segments indicating the center of mass and direction deduced for each
line segment in the image. If you try it on some test arrows, you'll
notice that the deduced direction will be affected by the weight of
a non-symmetrical arrowhead. This is bad. I'm going to rip out the
old LineSegment recognizer and replace it with something based on the
Hough transform.

The function WritePNG() is new code, and should not be considered
100% reliable. It might have bugs and/or leak.