Replace the Hough transform with something even better.
Third time's the charm for recognizing line segment coordinates.
The Hough transform was all right, but it still suffered from the
problem that an asymmetrical arrowhead (like the ones in
png-patterns/RotatingOogle/) would "weigh down" one side of the
shape and deflect the computed angle in the direction of the
extra weight.

My new simple solution is: Find the center of mass. Now find the
pixel that's farthest away from that center; assume it's one of
the endpoints of the line. (If the segment has an arrowhead,
this endpoint will be the non-arrow endpoint.)  Then, find the
pixel that's farthest away from that first endpoint; assume it's
the other endpoint of the line.  The line's direction is simply
the angle from endpoint 1 to endpoint 2, and we can even recompute
the "center" of the shape as the midpoint of that ideal line.
Detecting arrowheads is still done by the dumb method of detecting
any asymmetry in weight distribution.

This method produces slightly better results than Hough, and as
a bonus, it doesn't use a ton of trig functions!