Add "Matt's Modified Pretzel" to png-patterns/.
The paths in this pattern are a little distorted due to my desire
not to overlap any of the light-gray circles.

Notice that this pattern is sort of a super-turbo three-person
Morgan's Reel; it keeps the concept of one middle person "feeding"
the two end people in a three-count, even though you can also see
the pattern in the usual way as a five-count change-feeder where
the feeder is passing on 1 and 3.

    ./png2js --join=0 --count=1 ../png-patterns/MattsModifiedPretzel/*.png