Draw the clubs being held, as well as passes and selfs.
The Javascript for this feature is in a reasonably final form,
but the code in "png2js.cc" is just a placeholder. It currently
just loops over the timesteps and inserts a 1.7-beat hold every
beat, alternating hands. This works fine for all the currently
supported patterns because the only pass supported is the tramline
single, but it's obviously not how it ought to be done long-term.

In the long term, "png2js.cc" needs to generate holds based on
the causal diagram for the pattern (which it can infer based on
the destinations of the passes).