Add "Trefoil" and "Trefoil2" to png-patterns/.
Trefoil comes from the Santa Barbara jugglers:
http://www.youtube.com/watch?v=UgZSG-jGoG8

This is sort of a "lazy Havana". The feeder has the same
five passes as in Havana (albeit mirror-imaged), but when
the two feedees do-si-do, one of them becomes the next
feeder, rather than going to the "hang-out corner". The
count (just right-handed throws) is "PSSPSSSPSPSPPPPPPSSS".

"Trefoil2" is a messier variation on the theme, in which
the hangout corner becomes the next feeder. It's basically
Havana, but with the wrong side running. The count (just
right-handed throws) is "PSSSPSPSPSSSPPPPPPSS". I came up
with this variation independently, by messing up the diagrams
on my first attempt.